requirements.txt
googletrans==4.0.0rc1
mysql_connector_repackaged==0.3.1
